improved file_remover

+++ MockTraining/file_remover.py
@@ -1,4 +1,6 @@
 import os
+import os.path
 
 def rm(filename):
-    os.remove(filename)
+    if os.path.isfile(filename):
+        os.remove(filename)
